reference_validator.py, YAML_HotClasses.py, YAML_HotValidator.py, YAML_Enums.py, YAML_Hotfile.py
- code structured into files
- reference_validator: parses arguments, runs validator, returns output
- YAML_HotValidator: contains corresponding class without subclasses
- YAML_Hotfile: contains corresponding class without subclasses
- YAML_Enums: all enumeration types (YAML_Types, YAML_colours, YAML_tree_info)
- YAML_HotClasses: other classes previously in YAML_HotValidator (YAML_Env, YAML_Prop_Par, YAML_Reference, YAML_Resource)
